Move redirection helpers to new module
