+++ src/execute.c
@@ -28,6 +28,7 @@
 #include "util.h"
 #include "hash.h"
 #include "lexer.h"
+#include "redir.h"
 
 extern int last_status;
 extern int param_error;
@@ -39,7 +40,6 @@ int loop_depth = 0;
 int run_command_list(Command *cmds, const char *line);
 static int apply_temp_assignments(PipelineSegment *pipeline, int background,
                                   const char *line);
-void setup_redirections(PipelineSegment *seg);
 static int spawn_pipeline_segments(PipelineSegment *pipeline, int background,
                                    const char *line);
 static int run_pipeline_timed(PipelineSegment *pipeline, int background,
@@ -78,87 +78,6 @@ static int is_builtin_command(const char *name) {
             return 1;
     }
     return 0;
-}
-
-/* Apply redirections in the current process for builtin execution. */
-struct redir_save { int in, out, err; };
-static int apply_redirs_shell(PipelineSegment *seg, struct redir_save *sv) {
-    sv->in = sv->out = sv->err = -1;
-    if (seg->in_file) {
-        sv->in = dup(seg->in_fd);
-        int fd = open(seg->in_file, O_RDONLY);
-        if (fd < 0) {
-            perror(seg->in_file);
-            return -1;
-        }
-        if (seg->here_doc)
-            unlink(seg->in_file);
-        dup2(fd, seg->in_fd);
-        close(fd);
-    }
-
-    if (seg->out_file && seg->err_file && strcmp(seg->out_file, seg->err_file) == 0 &&
-        seg->append == seg->err_append) {
-        sv->out = dup(seg->out_fd);
-        sv->err = dup(STDERR_FILENO);
-        int fd = open_redirect(seg->out_file, seg->append, seg->force);
-        if (fd < 0) {
-            perror(seg->out_file);
-            return -1;
-        }
-        dup2(fd, STDOUT_FILENO);
-        dup2(fd, STDERR_FILENO);
-        close(fd);
-    } else {
-        if (seg->out_file) {
-            sv->out = dup(seg->out_fd);
-            int fd = open_redirect(seg->out_file, seg->append, seg->force);
-            if (fd < 0) {
-                perror(seg->out_file);
-                return -1;
-            }
-            dup2(fd, seg->out_fd);
-            close(fd);
-        }
-        if (seg->err_file) {
-            sv->err = dup(STDERR_FILENO);
-            int fd = open_redirect(seg->err_file, seg->err_append, 0);
-            if (fd < 0) {
-                perror(seg->err_file);
-                return -1;
-            }
-            dup2(fd, STDERR_FILENO);
-            close(fd);
-        }
-    }
-
-    if (seg->close_out) {
-        if (sv->out == -1)
-            sv->out = dup(seg->out_fd);
-        close(seg->out_fd);
-    } else if (seg->dup_out != -1) {
-        if (sv->out == -1)
-            sv->out = dup(seg->out_fd);
-        dup2(seg->dup_out, seg->out_fd);
-    }
-
-    if (seg->close_err) {
-        if (sv->err == -1)
-            sv->err = dup(STDERR_FILENO);
-        close(STDERR_FILENO);
-    } else if (seg->dup_err != -1) {
-        if (sv->err == -1)
-            sv->err = dup(STDERR_FILENO);
-        dup2(seg->dup_err, STDERR_FILENO);
-    }
-
-    return 0;
-}
-
-static void restore_redirs_shell(PipelineSegment *seg, struct redir_save *sv) {
-    if (sv->in != -1) { dup2(sv->in, seg->in_fd); close(sv->in); }
-    if (sv->out != -1) { dup2(sv->out, seg->out_fd); close(sv->out); }
-    if (sv->err != -1) { dup2(sv->err, STDERR_FILENO); close(sv->err); }
 }
 
 /* Execute a builtin or function with any redirections in the current shell. */
@@ -435,15 +354,6 @@ static PipelineSegment *copy_pipeline(PipelineSegment *src) {
     }
     return head;
 }
-/*
- * Duplicate the given descriptor onto another descriptor and close the
- * original.  Used internally when applying redirections.  Does not
- * modify last_status.
- */
-static void redirect_fd(int fd, int dest) {
-    dup2(fd, dest);
-    close(fd);
-}
 
 static char **parse_array_values(const char *val, int *count) {
     *count = 0;
@@ -677,64 +587,6 @@ static int apply_temp_assignments(PipelineSegment *pipeline, int background,
 
     return handled;
 }
-
-void setup_redirections(PipelineSegment *seg) {
-    if (seg->in_file) {
-        int fd = open(seg->in_file, O_RDONLY);
-        if (fd < 0) {
-            perror(seg->in_file);
-            exit(1);
-        }
-        if (seg->here_doc)
-            unlink(seg->in_file);
-        redirect_fd(fd, seg->in_fd);
-    }
-
-    if (seg->out_file && seg->err_file && strcmp(seg->out_file, seg->err_file) == 0 &&
-        seg->append == seg->err_append) {
-        int fd = open_redirect(seg->out_file, seg->append, seg->force);
-        if (fd < 0) {
-            perror(seg->out_file);
-            exit(1);
-        }
-        dup2(fd, STDOUT_FILENO);
-        dup2(fd, STDERR_FILENO);
-        close(fd);
-    } else {
-        if (seg->out_file) {
-            int fd = open_redirect(seg->out_file, seg->append, seg->force);
-            if (fd < 0) {
-                perror(seg->out_file);
-                exit(1);
-            }
-            redirect_fd(fd, seg->out_fd);
-        }
-        if (seg->err_file) {
-            int fd = open_redirect(seg->err_file, seg->err_append, 0);
-            if (fd < 0) {
-                perror(seg->err_file);
-                exit(1);
-            }
-            redirect_fd(fd, STDERR_FILENO);
-        }
-    }
-
-    if (seg->close_out) {
-        close(seg->out_fd);
-        if (seg->out_fd == STDERR_FILENO)
-            seg->close_err = 0;
-        if (seg->dup_err == seg->out_fd)
-            seg->dup_err = -1;
-    } else if (seg->dup_out != -1) {
-        dup2(seg->dup_out, seg->out_fd);
-    }
-
-    if (seg->close_err)
-        close(STDERR_FILENO);
-    else if (seg->dup_err != -1)
-        dup2(seg->dup_err, STDERR_FILENO);
-}
-
 
 /*
  * Fork and execute each segment of a pipeline, wiring up pipes between
